requirements.txt
discord.py==1.7.3
requests-html
beautifulsoup4==4.10.0
icalendar